requirements.txt
streamlit
openai
Pillow
requests

